Extract CLI test fakes
